Added a parzigzag_changeorientation helper function, which is currently triggered, if robot is blocked by wall N number of times
In future, it can also be triggered, if the robot has been walking
for a new predefined N number of steps. Or for other reasons.

As of now the check for quit parzigzag due to N number of orientation
changes having occured has been kept out of this new helper func
however if it gets used in many places in future, I may move it
into the helper func itself and use the return value from this
func, which will be the orientation change count check, to set
bContinue or so ... in parzigzag func.

Also enabled the default flow of parzigzag followed by wallfollow