IMPORTANT move_ext updated with default optional parameter markVisited
Now by default when ever move_ext is called, it will also update
the gVisited dictionary/map. So no need to call mark_as_visited
explicitly when moving the robot around.

NOTE: However if some moves are being done to check if something
is possible or not but actually the required logic is not carried
out at the new position, but only cross checked for future possibility
or some such reason and then the position is reverted back to the
original position. Then gVisited map will have wrong data, if one
is depending on it. IN SUCH CASES YOU WILL WANT TO EXPLICITLY
pass False to move_ext function when calling it.