Extended wallfollow to make it verygood at covering non wall cells using a optional argument.
By default if no argument is given wallfollow just does a wall
following logic as before.

However if True is passed as the argument to wallfollow, then the
logic which walks the corresponding row or col perpendicular to
the wall position (for each and every wall side position) where
the robot is currently, gets enabled. And this will potentially
help cover most if not all of the accessible floor area for the
robot.

So have moved the old default flow into a new function called
superwalk_ok1. And inturn the new extended wallfollow is used
as the new default flow by putting it within superwalk_good
and calling it by default.

TODO: Update/Extend wallfollow further such that if the robot
comes back to its starting point after doing a full wall follow
walking, then the walk is stopped.