WallFollow Updated to recognise repeat of wall walking
Now the WallFollow logic keeps track as to where it started to do
the wall walking and then recognises when it has come back to the
same position at the end of a given wall walking round. In turn
it can be configured to repeat the wall walking round N number of
times and then quit wall walking.

NOTE: By default this is configured to exit once 1 round of wall
walking is done. But can be changed by updating the global var
appropriately.

Also now it prints periodic update wrt wall walking