WallFollow walk either side of perpendicular row/col walk - Step2
If the perpendicular to wall edge row/col walk passes next to a
wall, then the WalkEitherSide logic would get out of sync. This
as been fixed now.

TODO: However once a wall is reached during the row/col walk logic,
it doesn't walk the left side line. This has to be fixed next