WallFollow Walk EitherSide of Perpendicular Row/Col Walk - Step 3
Instead of going one side on one way of the row/col walk and the
other side on the other way of the walk. Now both side walk occurs
when going one way and the other way just comes back to where it
started from.

This solves the problem of when the row/col walk hits a wall only
one of the side walks occurs and not the other side. Also it does
it by keeping the logic simple.