WallFollow walk eitherside during perpendicular row/colWalk - Step4
Made the WallFollow partially intelligent wrt when it triggers the
expensive walk eitherside during the perpendicular row/col walks.

It checks if the current X position at the wall side during wall
follow walk is near to a area which has been visited less (either
during a earlier ParZigZag or WallFollow walk or ... in future)
, in which case, it will enable iWalkEitherSide for the row/col
walk wrt that wall position.

Will add the same check for Y position also next.

TO help with same, added logic to identify the Width of the current
room/area along X and Y axis. This is done in find_visited_rowscols
function, which is called after a normal ParZigZag or WallFollow
run.