WallFollow walkEitherSide during perpendicularRowColWalk - Step5
Fixed 2 slightly Suttle bugs

a) If the Robot starts away from a wall, then there is logic to
hunt for a wall. However it was putting the Robot back into a
proper position for WallFollow walk, when in reality the logic
was expecting the Robot to be facing the wall, as part of identify
wall logic. So this has been fixed now.

b) The check for whether we have reached the starting position was
occuring at a slightly wrong place. It has been moved to the right
place now. (This would have occured, if the start position was next
to/after a opening in the wall which triggers a turn).

worlds/rect1_wallfollow04.wld would have triggered both the above
bugs.