WallFollow walkEitherSide during perpendicularRowColWalk - Step6
Updated/fixed near_lessvisitedarea such that when checking the
AlongX/YVisits dict, corresponding to the borders of the room/
area it doesn't try access beyond the border wrt AlongX/YVisits
which won't exist.

Also added the check for the YAxis also. <<TODO]

NOTE: Currently in this iteration of the implementation
purposefully the Visited Map/Dict is not used to optimize the
walking in the main run/logic of ParZigZag and WallFollow.
This is So that the logic is rather robust even if in-efficient.
It is used only wrt the 2nd run of WallFollow which is tasked
with seeking out/reach low visited areas.

My old Recursive (and also may be NonRecursive, don't remember all
the logics which I had implemented long time back, has it is few
years old now) logic does have mechanism to use the visited map
/dict to decide whether to do further processing or visit or ...
as required.