WallFollow WalkEitherSide for PerpendicularRowColWalk - Step7
Added cross check along Y axis also as part of near_lessvisited
area logic.

With this the basic WallFollowing and its basic flavours done.

Also this should cover most areas in a room with blockig objects
in between, except for the cases where a spiral is formed by
blocking objects, where the inner parts of the spiral area wont
be reached by the WallFollow logic unless we increase the iWalk
EitherSide depth argument (which is supported/provided by Wall
Follow logic, but not used by the superwalk logic currently.
In future where WallFollow is further refined to control the
iWalkEitherSide parameter in a more fine grained manner rather
than the current row or col under consideration (and not both),
it can be enabled, as it will be restricted to a more tighter
area. AND/OR a virtual boundry logic can be implemented and
then either this or my much older partly-dumb-partly-intelligent
recursiveTakeAllPathsForEachCell can be used for such less visited
/ hard to reach areas).