Exploratory yet efficient WallFollow - Step1
Previously WallFollow logic was made more thorough/exploratory by
increasing the iWalkEitherSide Depth/Level. However, this would
also lead to more number of wasted walks as any walk near to the
less visited area would have triggered this deeper (and equally
wasteful) SideWalks.

Instead now the SideWalk is made equally or rather more deeper/
exploratory by also making it intelligent. Now during the row/col
/side walks if the logic reaches a new cell, and if it is also at
the last leg of the side walk (rather a pure row/col walk with no
side walk is also the same) then new eitherside walk is triggered,
thus extending the sidewalk automatically. Also note that this
auto extending sidewalk will keep occuring as long as it lands on
new not visited cells.

Because of above, the depth of WalkEitherSide for the 2nd wallfollow
wrt superwalk_good has been reduced from 2 to 1.

There is one more key change required for the new intelligently
efficient exploratory wallfollow to work fully properly, which I
will be commiting next. To be continued ...

Also to help with this and potentially other new usecases in future
the Visited Map/Dictionary has been changed from a binary map to a
integer/counter based one. This was required now itself because
the check for already visited or not is being done after the cell
as already been reached AND PARALLELY move_ext by default sets the
visited map as it moves the robot, to help make the default case for
visited map automatic, with no additional effort required.

NOTE: If required the row/col/side walk logic could have asked its
move_ext call to not set the visited map by default, and then after
r/c/s walk logic actually checks as to whether the current cell it
is on is new cell or not, set the visited map explicitly.