Intelligently Efficient Exploratory row/col/eitherside walk -Step2
To make it possible to allow a initial limited mode wallfollow with
row/col walk to map out easily reached areas in the map. And then
follow it up with a Exploratory wallfollow with row/col/eitherside
walk. A new bExplore argument as been added to WallFollow and
associated logic (like row/col/eitherside walk).

In turn the superwalk_good which tries to achieve the above balance
and logic uses the bExplore suitabley wrt the two wallfollow calls
that it makes.