WallFollow ProcessVisited periodically
When ever Explore EitherSideWalking is triggered in bExplore mode
because we have visited a new cell, a flag is set. In turn as the
WallFollow keeps moving along the wall, it periodically checks
for this processVisited flag and inturn calls find_visited_rowscols
so that the WallFollow can avoid detailed walking around low visited
areas, if the area was covered as part of the current walk till now.